+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added automatic line break for comments on Excel
 - Limit the number of comments on a sheet at a little less than 1 048 576
 - It's now possible to download comments from a single video
+- Added a new parameter to sort comments from oldest to newest or newest to oldest
 - Fix a bug where the application window isn't in the monitors coordinates
 - Fix a bug where the previous old save is used if there isn't any old save currently
+- Fix a bug where the upload_date is not retrieved from old saves
 
 ## [1.0.0] - 2025-03-23
 - First version of the application with all the essential features.